requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
APScheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
